+++ packages/nx/src/utils/errors.ts
@@ -1,4 +1,4 @@
-import { stringUtils } from '@nx/workspace';
+import * as nxStringUtils from '@nx/devkit/src/utils/string-utils';
 import { supportedFrameworks } from './general';
 
 export function missingArgument(argName: string, description: string = '', example: string = '') {
@@ -6,7 +6,7 @@ export function missingArgument(argName: string, description: string = '', examp
 }
 
 export function unsupportedFrameworkError(framework: string) {
-  return `${framework} is currently not a supported framework. Supported at the moment: ${supportedFrameworks.map((f) => stringUtils.capitalize(f))}. Please request support for this framework if you'd like and/or submit a PR which we would greatly appreciate.`;
+  return `${framework} is currently not a supported framework. Supported at the moment: ${supportedFrameworks.map((f) => nxStringUtils.capitalize(f))}. Please request support for this framework if you'd like and/or submit a PR which we would greatly appreciate.`;
 }
 
 export function generateOptionError(type: string, missingFeature?: boolean) {