+++ packages/nx/src/utils/versions.ts
@@ -1,18 +1,18 @@
-export const nxVersion = '^16.3.0';
+export const nxVersion = '^17.0.0';
 
-export const nsNxPluginVersion = '^16.3.0';
-export const nsCoreVersion = '~8.5.0';
-export const nsTypesVersion = '~8.5.0';
+export const nsNxPluginVersion = '^17.0.0';
+export const nsCoreVersion = '~8.6.0';
+export const nsTypesVersion = '~8.6.0';
 export const nsThemeVersion = '~1.0.4';
 export const nsWebpackVersion = '~5.0.0';
 export const sassVersion = '^1.32.0';
-export const nsIOSRuntimeVersion = '~8.5.0';
-export const nsAndroidRuntimeVersion = '~8.5.0';
+export const nsIOSRuntimeVersion = '~8.6.0';
+export const nsAndroidRuntimeVersion = '~8.6.0';
 
 // Frontend frameworks
-export const angularVersion = '^16.0.0';
-export const nsAngularVersion = '^16.0.0';
-export const nsNgToolsVersion = '^16.0.0';
+export const angularVersion = '^17.0.0';
+export const nsAngularVersion = '^17.0.0';
+export const nsNgToolsVersion = '^17.0.0';
 export const rxjsVersion = '~7.8.0';
-export const zonejsVersion = '^0.13.0';
-export const typescriptVersion = '~5.0.0';
+export const zonejsVersion = '^0.14.0';
+export const typescriptVersion = '~5.2.0';